requirements.txt
openai>=1.0.0
tqdm>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
